// .drizzle/schema.ts
/** Automatically generated by @palmares/drizzle-engine on 2024-12-01T19:57:35.258Z */ 

import * as d from '@palmares/drizzle-engine/sqlite-core';
import * as drzl from '@palmares/drizzle-engine/drizzle';

export const User = d.sqliteTable('user', {
  id: d.integer('id', { mode: 'number' }).primaryKey({ autoIncrement: true }).unique(),
  firstName: d.text('first_name', { length: 255 }).notNull(),
  lastName: d.text('last_name', { length: 255 }).notNull(),
  email: d.text('email').notNull().unique(),
  isActive: d.integer('is_active', { mode: 'boolean' }).notNull()
}, (table) => ({
  idIdx: d.uniqueIndex('user_id_idx').on(table.id)
}));

export const Analytics = d.sqliteTable('analytics', {
  id: d.integer('id', { mode: 'number' }).primaryKey({ autoIncrement: true }).unique(),
  ip: d.text('ip').notNull(),
  userAgent: d.text('user_agent').notNull(),
  referer: d.text('referer').notNull(),
  shortId: d.integer('short_id', { mode: 'number' }).notNull().references((): d.AnySQLiteColumn => Shorter.id)
}, (table) => ({
  idIdx: d.uniqueIndex('analytics_id_idx').on(table.id)
}));

export const Shorter = d.sqliteTable('shorter', {
  id: d.integer('id', { mode: 'number' }).primaryKey({ autoIncrement: true }).unique(),
  short: d.text('short', { length: 255 }).notNull(),
  long: d.text('long').notNull(),
  isActive: d.integer('is_active', { mode: 'boolean' }).notNull(),
  userId: d.integer('user_id', { mode: 'number' }).notNull().references((): d.AnySQLiteColumn => User.id)
}, (table) => ({
  idIdx: d.uniqueIndex('shorter_id_idx').on(table.id)
}));

export const AnalyticsRelations = drzl.relations(Analytics, (args) => ({
  analyticsShorterLink: args.one(Shorter, {
    fields: [Analytics.shortId],
    references: [Shorter.id]
  })
}));

export const ShorterRelations = drzl.relations(Shorter, (args) => ({
  analyticsRecords: args.many(Analytics),
  userLink: args.one(User, {
    fields: [Shorter.userId],
    references: [User.id]
  })
}));

export const UserRelations = drzl.relations(User, (args) => ({
  shorterLinks: args.many(Shorter)
}));